requirements.txt
fastapi
pytube
uvicorn
openai
python-dotenv
